Allocate Mat4 buffer as StaticArray
This allows the buffer contents to also be placed on the stack.
In its current form, it will allocate the buffer in the heap while
placing the struct containing a reference to it on the stack.

I've also made it include the `Indexable` module which makes it a
more flexible array-like object.